# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for the PMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API"
        }
    ],
    "paths": {
...
        "/projects/{project}/epics": {
            "get": {
                "tags": [
                    "Epics"
                ],
                "summary": "List epics for a project",
                "description": "Returns a list of all epics in the specified project.",
                "operationId": "listEpics",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Epics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Epics"
                ],
                "summary": "Create a new epic",
                "description": "Creates a new epic in the specified project.",
                "operationId": "createEpic",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the epic",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the epic",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the epic",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "description": "Due date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "assignee_id": {
                                        "description": "ID of the assignee user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Epic created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/epics/{epic}": {
            "get": {
                "tags": [
                    "Epics"
                ],
                "summary": "Get epic details",
                "description": "Returns details of a specific epic.",
                "operationId": "getEpic",
                "parameters": [
                    {
                        "name": "epic",
                        "in": "path",
                        "description": "The ID of the epic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Epic retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Epics"
                ],
                "summary": "Update an epic",
                "description": "Updates an existing epic.",
                "operationId": "updateEpic",
                "parameters": [
                    {
                        "name": "epic",
                        "in": "path",
                        "description": "The ID of the epic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The title of the epic",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the epic",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the epic",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "description": "Due date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "assignee_id": {
                                        "description": "ID of the assignee user",
                                        "type": "integer"
                                    },
                                    "position": {
                                        "description": "Position in list",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Epic updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Epics"
                ],
                "summary": "Delete an epic",
                "description": "Deletes an existing epic.",
                "operationId": "deleteEpic",
                "parameters": [
                    {
                        "name": "epic",
                        "in": "path",
                        "description": "The ID of the epic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Epic deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}